* CmdExec.cc: fixed setting of cmdline in new jobs. 	* Torrent.cc, lftp_rl.c, xstring.cc: fixed warnings 	  (patch from Openwall linux, Alexander Peslyak). 	* complete.cc: use CmdExec::is_space for parsing. 	* xstring.h: fixed compilation with non-gcc and old gcc compilers 	  (patch from Openwall linux, Alexander Peslyak).

+++ src/xstring.cc
@@ -379,7 +379,7 @@ const char *xstring::dump_to(xstring& buf) const
 	    buf.appendf("%02X",(unsigned char)*s++);
 	 buf.append('>');
       } else {
-	 buf.appendf("<long binary, %d bytes>",length());
+	 buf.appendf("<long binary, %d bytes>",(int)length());
       }
    } else {
       while(len>0) {
